+++ aio/content/examples/aot-compiler/rollup-config.js
@@ -1,8 +1,7 @@
 // #docregion
-import rollup      from 'rollup'
-import nodeResolve from 'rollup-plugin-node-resolve'
+import nodeResolve from 'rollup-plugin-node-resolve';
 import commonjs    from 'rollup-plugin-commonjs';
-import uglify      from 'rollup-plugin-uglify'
+import uglify      from 'rollup-plugin-uglify';
 
 // #docregion config
 export default {
@@ -30,5 +29,5 @@ export default {
       uglify()
       // #enddocregion uglify
   ]
-}
+};
 // #enddocregion config